+++ server/db/log.go
@@ -12,8 +12,8 @@ import (
 
 const slowThreshold = 200 * time.Millisecond // same as GORM
 
-// FieldLogger is the same as logrus.FieldLogger, but with WithContext that was
-// accidentally omitted.
+// FieldLogger is the same as log.FieldLogger, but with an additional
+// WithContext method.
 type FieldLogger interface {
 	log.FieldLogger
 	WithContext(ctx context.Context) *log.Entry
@@ -24,22 +24,27 @@ type Logger struct {
 	Logger FieldLogger
 }
 
+// LogMode implements GORM's logger.Interface and does nothing.
 func (l Logger) LogMode(_ gormlogger.LogLevel) gormlogger.Interface {
 	return l // ignore log mode change requests from GORM :)
 }
 
+// Error implements GORM's logger.Interface.
 func (l Logger) Error(ctx context.Context, s string, i ...interface{}) {
 	l.Logger.WithContext(ctx).Errorf(s, i...)
 }
 
+// Warn implements GORM's logger.Interface.
 func (l Logger) Warn(ctx context.Context, s string, i ...interface{}) {
 	l.Logger.WithContext(ctx).Warnf(s, i...)
 }
 
+// Info implements GORM's logger.Interface.
 func (l Logger) Info(ctx context.Context, s string, i ...interface{}) {
 	l.Logger.WithContext(ctx).Infof(s, i...)
 }
 
+// Trace implements GORM's logger.Interface.
 func (l Logger) Trace(ctx context.Context, begin time.Time, fc func() (string, int64), err error) {
 	elapsed := time.Since(begin)
