+++ server/db/meetup.go
@@ -7,6 +7,7 @@ import (
 	"gorm.io/gorm"
 )
 
+// Meetup is the database model for a meetup.
 type Meetup struct {
 	gorm.Model
 	Title             string
@@ -23,6 +24,7 @@ type Meetup struct {
 	Cancelled         bool
 }
 
+// IDString returns the meetup's ID as a string.
 func (m *Meetup) IDString() string {
 	if m == nil || m.ID == 0 {
 		return ""
@@ -30,15 +32,20 @@ func (m *Meetup) IDString() string {
 	return strconv.FormatUint(uint64(m.ID), 10)
 }
 
+// MeetupIDFromString reverses Meetup.IDString.
 func MeetupIDFromString(s string) (uint, error) {
 	id, err := strconv.ParseUint(s, 10, 0)
 	return uint(id), err
 }
 
+// Coordinates represent WGS 84 coordinates of the earth. It is not a database
+// model but rather stored as a part of MeetupLocation.
 type Coordinates struct {
 	Lat, Lon *float64
 }
 
+// MeetupLocation represents where a meetup could be held. It is not a database
+// model but rather stored as a part of Meetup.
 type MeetupLocation struct {
 	Coordinates
 	URL  string
