+++ server/db/user.go
@@ -6,6 +6,7 @@ import (
 	"gorm.io/gorm"
 )
 
+// User is the database model for users.
 type User struct {
 	gorm.Model
 	Email           string `gorm:"uniqueIndex"`
@@ -21,6 +22,7 @@ type User struct {
 	PendingApproval []*Meetup   `gorm:"many2many:meetup_user_pending;"`
 }
 
+// IDString returns the user's ID, represented as a string.
 func (u *User) IDString() string {
 	if u == nil || u.ID == 0 {
 		return ""
@@ -28,6 +30,7 @@ func (u *User) IDString() string {
 	return strconv.FormatUint(uint64(u.ID), 10)
 }
 
+// UserIDFromString reverses User.IDString.
 func UserIDFromString(s string) (uint, error) {
 	id, err := strconv.ParseUint(s, 10, 0)
 	return uint(id), err
