+++ server/impl/log.go
@@ -2,13 +2,20 @@ package impl
 
 import log "github.com/sirupsen/logrus"
 
+// RequestLogHook is a log.Hook implementation that adds to the log.Entry
+// information gleaned from the entry context, such as the HTTP request (using
+// RequestFromContext) and the user session (using SessionFromContext).
 type RequestLogHook struct{}
 
+var _ log.Hook = RequestLogHook{}
+
+// Levels returns log.AllLevels and implements log.Hook.
 func (r RequestLogHook) Levels() []log.Level { return log.AllLevels }
 
+// Fire implements log.Hook.
 func (r RequestLogHook) Fire(entry *log.Entry) error {
 	if ctx := entry.Context; ctx != nil {
-		if r := RequestFromSession(ctx); r != nil {
+		if r := RequestFromContext(ctx); r != nil {
 			tentativeAddField(entry.Data, "method", r.Method)
 			tentativeAddField(entry.Data, "url", r.URL.String())
 		}
