+++ server/impl/nonce/nonce_test.go
@@ -23,7 +23,7 @@ func TestGenerator_NewState_Concurrency(t *testing.T) {
 		go func() {
 			defer wg.Done()
 			for j := 0; j < repetition; j++ {
-				g.NewState(stateLen)
+				g.NewBase64(stateLen)
 			}
 		}()
 	}
